requirements.txt
flask
google-api-python-client
google-auth
google-auth-oauthlib
google-auth-httplib2
matplotlib
pandas
openpyxl
reportlab
gunicorn